type compare

+++ hunt.cpp
@@ -130,7 +130,7 @@ void read_map(crew &crew_in) {
     getline(cin, s);
     if(first == 'L') {
         while(getline(cin, s)) {
-            if(s.length() == 5) {
+            if((int)s.length() == 5) {
                 
                 int row = s[0] - '0';
                 int col = s[2] - '0';
